refactor: consolidate enum types from db-schema to @repo/schemas
- Remove 7 duplicate type definitions from db-schema tables
  (AgentRuntime, AgentRunStatus, JobStatus, JobType, RuleSeverity,
  RuleCategory, RuleScriptLanguage, SpanType, SpanStatus)
- Import types from @repo/schemas in db-schema tables
- Update all consumers to import enum types from @repo/schemas
- Add operation_run to JobTypeSchema (was in db-schema but missing)
- Move Job/PipelineRun schemas into job/ subdirectory
- Remove @repo/db-schema dep from @repo/obs (no longer needed)
- Add @repo/schemas dep to @repo/models and @repo/obs
- Add 'no re-export' rule to do-not.instructions.md
