+++ packages/models/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@repo/db": "workspace:*",
     "@repo/db-schema": "workspace:*",
+    "@repo/schemas": "workspace:*",
     "drizzle-orm": "^0.45.1"
   },
   "devDependencies": {
