+++ packages/obs/package.json
@@ -13,7 +13,6 @@
     "quality": "bun run check-types && bun run lint"
   },
   "dependencies": {
-    "@repo/db-schema": "workspace:*",
     "@repo/models": "workspace:*",
     "@repo/schemas": "workspace:*"
   },
