+++ packages/schemas/src/index.ts
@@ -17,11 +17,7 @@ export * from "./ExecutorConfigSchema";
 export * from "./ExecutorTypeSchema";
 export * from "./GithubProjectSchema";
 export * from "./InputPortSchema";
-export * from "./JobResultSchema";
-export * from "./JobSchema";
-export * from "./JobStatusSchema";
-export * from "./JobTypeSchema";
-export * from "./JobTraceSchema";
+export * from "./job";
 export * from "./LocalConnectionSchema";
 export * from "./LogLevelSchema";
 export * from "./meta";
@@ -29,7 +25,6 @@ export * from "./ObjectTypeSchema";
 export * from "./OperationConfigSchema";
 export * from "./OperationSchema";
 export * from "./OutputItemSchema";
-export * from "./PipelineRunSchema";
 export * from "./PortKindSchema";
 export * from "./RecipeSchema";
 export * from "./RefinementRunSchema";
