+++ packages/schemas/src/job/PipelineRunSchema.ts
@@ -1,6 +1,6 @@
 import { z } from "zod/v4";
 import { JobResultSchema } from "./JobResultSchema";
-import { MetaSchema } from "./meta";
+import { MetaSchema } from "../meta";
 
 export const PipelineRunSchema = z.object({
   id: z.string(),
@@ -9,7 +9,6 @@ export const PipelineRunSchema = z.object({
   inputPath: z.string().nullable(),
   logs: z.array(z.string()),
   result: JobResultSchema.nullable(),
-  tmuxSessionName: z.string().nullable(),
   meta: MetaSchema.optional(),
 });
 export type PipelineRun = z.infer<typeof PipelineRunSchema>;
